Switch to type erased method. Const generics become unconstrained in typed APIs.
