+++ src/for_each/for_each_3arity.rs
