+++ src/lib.rs
@@ -152,7 +152,6 @@ pub mod prelude {
         operations::{
             //for_each::*,
             AeryQueryExt,
-            ControlFlow,
             EdgeSide,
             FoldBreadth,
             Join,
